@@ -3,7 +3,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 void main() {
-  runApp(ProviderScope(child: MemoryExample()));
+  runApp(const ProviderScope(child: MemoryExample()));
 }
 
 final countProvider = StateNotifierProvider((_) => CountProvider(1));
@@ -11,6 +11,8 @@ final countProvider = StateNotifierProvider((_) => CountProvider(1));
 final currentCount = Provider((ref) => ref.watch(countProvider));
 
 class MemoryExample extends HookConsumerWidget {
+  const MemoryExample({super.key});
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final notifier = ref.watch(countProvider.notifier);
@@ -21,13 +23,13 @@ class MemoryExample extends HookConsumerWidget {
       ),
       home: Scaffold(
         appBar: AppBar(
-          title: Text("Riverpod Animation Example"),
+          title: const Text("Riverpod Animation Example"),
         ),
         body: Column(
           children: [
-            AnimatedWidget(),
+            const AnimatedWidget(),
             ElevatedButton(
-              child: Text("Increment"),
+              child: const Text("Increment"),
               onPressed: () {
                 notifier.increment();
               },
@@ -42,34 +44,35 @@ class MemoryExample extends HookConsumerWidget {
 class AnimatedWidget extends HookConsumerWidget {
   final Duration duration = const Duration(milliseconds: 1000);
 
+  const AnimatedWidget({super.key});
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final _controller = useAnimationController(duration: duration);
-    _controller.addStatusListener((status) {
+    final controller = useAnimationController(duration: duration);
+    controller.addStatusListener((status) {
       if (status == AnimationStatus.completed) {
-        _controller.reverse();
+        controller.reverse();
       }
     });
 
-    var _count = ref.watch(currentCount);
-    useValueChanged(_count, (_, __) async {
-      _controller.forward();
+    var count = ref.watch(currentCount);
+    useValueChanged(count, (_, __) async {
+      controller.forward();
     });
 
     return AnimatedBuilder(
-        animation: _controller,
+        animation: controller,
         builder: (context, child) {
           return Center(
             child: Container(
-              margin: EdgeInsets.all(10),
-              decoration: BoxDecoration(
-                  color: Colors.green[300]!.withOpacity(_controller.value),
-                  shape: BoxShape.circle),
+              margin: const EdgeInsets.all(10),
+              decoration:
+                  BoxDecoration(color: Colors.green[300]!.withOpacity(controller.value), shape: BoxShape.circle),
               height: 200,
               width: 200,
               child: Center(
                   child: Text(
-                _count.toString(),
+                count.toString(),
                 textScaleFactor: 5,
               )),
             ),
